fix(agent): preserve bounded resume semantics
