+++ packages/agent/test/system-record-reconcile-v1.test.ts
@@ -1,28 +1,15 @@
 import { describe, expect, it, vi } from 'vitest';
 
-import { decodeOpaqueKaBundleV1 } from '@origintrail-official/dkg-core';
 import {
   SYSTEM_RECORD_CONTINUATION_TIMEOUT_MS,
-  SYSTEM_RECORD_MAX_CONTINUATION_CLOSURE_WIRE_BYTES,
   SYSTEM_RECORD_MAX_CONTINUATION_SLICES,
-  SYSTEM_RECORD_MAX_SLICE_WIRE_BYTES,
   buildSystemRecordInventoryTreeV1,
-  buildSystemRecordProviderSignatureMessageV1,
-  computeSystemRecordRootDescriptorDigestV1,
   computeSystemRecordStableKeyHashV1,
-  parseCanonicalSignedSystemRecordRootDescriptorEnvelopeV1,
-  type SignedSystemRecordRootDescriptorEnvelopeV1,
   type SystemRecordInventoryRowV1,
 } from '@origintrail-official/dkg-core/system-record-v1';
 
-import { parseNQuads } from '../src/dkg-agent-utils.js';
 import type { AgentProfileAdmittedSliceContextV1 } from '../src/system-records/admitted-slice-context-v1.js';
-import type { SystemRecordArtifactRepositoryV1 } from '../src/system-records/artifact-v1.js';
-import {
-  createAgentProfileReceiverV1,
-  type AgentProfilePreparedActiveV1,
-  type AgentProfileReceiverV1,
-} from '../src/system-records/receiver-v1.js';
+import type { AgentProfileReceiverV1 } from '../src/system-records/receiver-v1.js';
 import {
   createAgentProfileReconcilerV1,
   type AgentProfileReconcileAdmissionV1,
@@ -36,309 +23,18 @@ import type {
   SystemRecordExactFetchResultV1,
 } from '../src/system-records/requester-v1.js';
 import {
-  createFixtureAgentProfileProducerV1,
-  NETWORK,
-  produce,
-  producerFixture,
-  PRODUCER_FIXTURE_NOW_MS,
-  DEPLOYMENT,
-} from './support/agent-profile-producer-v1-fixture.js';
-
-const TEST_PEER_IDS = Object.freeze([
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwkf',
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwkg',
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwkh',
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwki',
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwkj',
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwkk',
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwkm',
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwkn',
-  '12D3KooW9pNAk8aiBuGVQtWRdbkLmo5qVL3e2h5UxbN2Nz9ttwko',
-]);
+  TEST_PEER_IDS,
+  admissionGate,
+  byteAdmission,
+  publishedFixture,
+  receiver,
+  receiverWithPreparation,
+  resignRootTotalRows,
+  signRootDescriptor,
+} from './support/agent-profile-reconcile-v1-fixture.js';
+import { NETWORK } from './support/agent-profile-producer-v1-fixture.js';
 
 describe('agent-profile System Record reconciler V1', () => {
-  it('routes inventory and receiver closure fetches through one admitted transport slice at a time', async () => {
-    const fixture = await publishedFixture();
-    const released: ReturnType<typeof vi.fn>[] = [];
-    const fetchExact = vi.fn(async (
-      _providerId: string,
-      lookup: SystemRecordExactArtifactLookupV1,
-      signal: AbortSignal,
-    ): Promise<SystemRecordExactFetchResultV1> => {
-      const artifact = await fixture.store.resolve(lookup, signal);
-      if (artifact === null) return Object.freeze({ outcome: 'not-found', wireBytes: 5 });
-      const release = vi.fn();
-      released.push(release);
-      return Object.freeze({
-        outcome: 'ok',
-        lease: Object.freeze({
-          artifact,
-          wireBytes: 4 + 128 + artifact.canonicalBytes.byteLength,
-          release,
-        }),
-      });
-    });
-    const transport = createAgentProfileReconcileTransportV1({
-      listProviderIds: () => ['provider-a'],
-      fetchExact,
-      controlAdmission: byteAdmission(),
-    });
-    const consumeCandidate = vi.fn(async () => ({
-      outcome: 'applied' as const,
-      stateRevision: '1',
-      appliedStateDigest: `0x${'aa'.repeat(32)}`,
-    }));
-    const reconciler = await createAgentProfileReconcilerV1({
-      networkId: NETWORK,
-      rootEnvelope: fixture.rootEnvelope,
-      providerPeerPublicKey: fixture.peerSigner.publicKey,
-      admission: admissionGate(),
-      transport,
-      receiver: receiver(fixture.store, consumeCandidate),
-    });
-
-    const inventoryResult = await reconciler.advance(new AbortController().signal);
-    expect(inventoryResult).toMatchObject({
-      status: 'paused',
-      phase: 'records',
-      inventoryRequests: 1,
-      closureWireBytes: 0,
-    });
-    const wireBytesAfterInventory = transport.stats().wireBytes;
-    expect(transport.stats()).toMatchObject({ activeSlice: 0, requests: 1 });
-    expect(released).toHaveLength(1);
-    expect(released[0]).toHaveBeenCalledTimes(1);
-
-    const recordsResult = await reconciler.advance(new AbortController().signal);
-    expect(recordsResult).toMatchObject({
-      status: 'complete',
-      processedRows: 1,
-    });
-    const closureWireBytes = transport.stats().wireBytes - wireBytesAfterInventory;
-    expect(recordsResult.closureWireBytes).toBe(closureWireBytes);
-    expect(reconciler.stats().closureWireBytes).toBe(closureWireBytes);
-    expect(consumeCandidate).toHaveBeenCalledTimes(1);
-    expect(fetchExact.mock.calls.every(([, lookup]) => lookup.type !== 'root')).toBe(true);
-    expect(fetchExact.mock.calls.length).toBeGreaterThanOrEqual(4);
-    expect(released.every((release) => release.mock.calls.length === 1)).toBe(true);
-    expect(transport.stats()).toMatchObject({
-      activeSlice: 0,
-      requests: fetchExact.mock.calls.length,
-      negativeMemoEntries: 0,
-    });
-  });
-
-  it('accounts every failed and successful provider attempt in inventory continuation bytes', async () => {
-    const fixture = await publishedFixture();
-    let successfulInventoryBytes = 0;
-    const fetchExact = vi.fn(async (
-      providerId: string,
-      lookup: SystemRecordExactArtifactLookupV1,
-      signal: AbortSignal,
-    ): Promise<SystemRecordExactFetchResultV1> => {
-      if (providerId === 'provider-a' && lookup.type === 'inventory-object') {
-        return Object.freeze({ outcome: 'not-found', wireBytes: 1_000 });
-      }
-      const artifact = await fixture.store.resolve(lookup, signal);
-      if (artifact === null) return Object.freeze({ outcome: 'not-found', wireBytes: 1 });
-      successfulInventoryBytes = 4 + 128 + artifact.canonicalBytes.byteLength;
-      return Object.freeze({
-        outcome: 'ok',
-        lease: Object.freeze({
-          artifact,
-          wireBytes: successfulInventoryBytes,
-          release: vi.fn(),
-        }),
-      });
-    });
-    const transport = createAgentProfileReconcileTransportV1({
-      listProviderIds: () => ['provider-a', 'provider-b'],
-      fetchExact,
-      controlAdmission: byteAdmission(),
-    });
-    const reconciler = await createAgentProfileReconcilerV1({
-      networkId: NETWORK,
-      rootEnvelope: fixture.rootEnvelope,
-      providerPeerPublicKey: fixture.peerSigner.publicKey,
-      admission: admissionGate(),
-      transport,
-      receiver: receiver(fixture.store, vi.fn()),
-    });
-    const result = await reconciler.advance(new AbortController().signal);
-    expect(result).toMatchObject({
-      status: 'paused',
-      phase: 'records',
-      inventoryRequests: 1,
-      inventoryWireBytes: 1_000 + successfulInventoryBytes,
-    });
-    expect(reconciler.stats()).toMatchObject({
-      inventoryRequests: 1,
-      inventoryWireBytes: 1_000 + successfulInventoryBytes,
-    });
-    expect(fetchExact.mock.calls.map(([providerId]) => providerId))
-      .toEqual(['provider-a', 'provider-b']);
-  });
-
-  it('pauses and accounts a retryable transport failure during closure preparation', async () => {
-    const fixture = await publishedFixture();
-    let closureBusy = true;
-    const fetchExact = vi.fn(async (
-      _providerId: string,
-      lookup: SystemRecordExactArtifactLookupV1,
-      signal: AbortSignal,
-    ): Promise<SystemRecordExactFetchResultV1> => {
-      if (lookup.type !== 'inventory-object' && closureBusy) {
-        return Object.freeze({ outcome: 'remote-busy', wireBytes: 13 });
-      }
-      const artifact = await fixture.store.resolve(lookup, signal);
-      if (artifact === null) return Object.freeze({ outcome: 'not-found', wireBytes: 1 });
-      return Object.freeze({
-        outcome: 'ok',
-        lease: Object.freeze({
-          artifact,
-          wireBytes: 4 + 128 + artifact.canonicalBytes.byteLength,
-          release: vi.fn(),
-        }),
-      });
-    });
-    const transport = createAgentProfileReconcileTransportV1({
-      listProviderIds: () => ['provider-a'],
-      fetchExact,
-      controlAdmission: byteAdmission(),
-    });
-    const reconciler = await createAgentProfileReconcilerV1({
-      networkId: NETWORK,
-      rootEnvelope: fixture.rootEnvelope,
-      providerPeerPublicKey: fixture.peerSigner.publicKey,
-      admission: admissionGate(),
-      transport,
-      receiver: receiver(fixture.store, async () => ({
-        outcome: 'applied',
-        stateRevision: '1',
-        appliedStateDigest: `0x${'aa'.repeat(32)}`,
-      })),
-    });
-
-    await reconciler.advance(new AbortController().signal);
-    await expect(reconciler.advance(new AbortController().signal)).resolves.toMatchObject({
-      status: 'paused',
-      phase: 'records',
-      closureWireBytes: 13,
-      pendingRows: 1,
-    });
-    expect(reconciler.stats()).toMatchObject({
-      advances: 1,
-      closureWireBytes: 13,
-      pendingRows: 1,
-    });
-
-    closureBusy = false;
-    await expect(reconciler.advance(new AbortController().signal)).resolves.toMatchObject({
-      status: 'complete',
-      processedRows: 1,
-    });
-  });
-
-  it('stops a continuation at the dedicated closure wire-byte budget', async () => {
-    const fixture = await publishedFixture();
-    const fetchExact = vi.fn(async (
-      _providerId: string,
-      lookup: SystemRecordExactArtifactLookupV1,
-      signal: AbortSignal,
-    ): Promise<SystemRecordExactFetchResultV1> => {
-      if (lookup.type !== 'inventory-object') {
-        return Object.freeze({
-          outcome: 'remote-error',
-          wireBytes: SYSTEM_RECORD_MAX_SLICE_WIRE_BYTES / 2,
-        });
-      }
-      const artifact = await fixture.store.resolve(lookup, signal);
-      if (artifact === null) return Object.freeze({ outcome: 'not-found', wireBytes: 1 });
-      return Object.freeze({
-        outcome: 'ok',
-        lease: Object.freeze({
-          artifact,
-          wireBytes: 4 + 128 + artifact.canonicalBytes.byteLength,
-          release: vi.fn(),
-        }),
-      });
-    });
-    const transport = createAgentProfileReconcileTransportV1({
-      listProviderIds: () => ['provider-a', 'provider-b'],
-      fetchExact,
-      controlAdmission: byteAdmission(),
-    });
-    const apply = vi.fn(async () => ({
-      outcome: 'applied' as const,
-      stateRevision: '1',
-      appliedStateDigest: `0x${'aa'.repeat(32)}`,
-    }));
-    const artifacts = fixture.store;
-    const prepareActive: AgentProfileReceiverV1['prepareActive'] = async (
-      row,
-      source,
-      signal,
-    ) => {
-        await source.resolve(Object.freeze({
-          type: 'object',
-          objectKind: 'agent-profile-head',
-          objectDigest: row.headDigest,
-        }), signal);
-        return Object.freeze({ apply });
-      };
-    const receiverWithRemotePreparation: AgentProfileReceiverV1 = Object.freeze({
-      artifacts,
-      openPreparation(row) {
-        return Object.freeze({
-          prepare: (source, signal) => prepareActive(row, source, signal),
-          release: () => undefined,
-        });
-      },
-      prepareActive,
-      async receiveActive(row, admittedContext, signal) {
-        const prepared = await prepareActive(row, artifacts, signal);
-        return prepared.apply(admittedContext, signal);
-      },
-    });
-    const reconciler = await createAgentProfileReconcilerV1({
-      networkId: NETWORK,
-      rootEnvelope: fixture.rootEnvelope,
-      providerPeerPublicKey: fixture.peerSigner.publicKey,
-      admission: admissionGate(),
-      transport,
-      receiver: receiverWithRemotePreparation,
-    });
-
-    await reconciler.advance(new AbortController().signal);
-    const closureSlices = SYSTEM_RECORD_MAX_CONTINUATION_CLOSURE_WIRE_BYTES
-      / SYSTEM_RECORD_MAX_SLICE_WIRE_BYTES;
-    for (let slice = 0; slice < closureSlices; slice += 1) {
-      const result = await reconciler.advance(new AbortController().signal);
-      if (result.status !== 'paused') {
-        throw new Error(`unexpected slice ${slice}: ${JSON.stringify({
-          result,
-          stats: reconciler.stats(),
-        })}`);
-      }
-      expect(result).toMatchObject({
-        status: 'paused',
-        phase: 'records',
-        closureWireBytes: SYSTEM_RECORD_MAX_SLICE_WIRE_BYTES,
-      });
-    }
-    expect(reconciler.stats()).toMatchObject({
-      closureWireBytes: SYSTEM_RECORD_MAX_CONTINUATION_CLOSURE_WIRE_BYTES,
-      processedRows: 0,
-      pendingRows: 1,
-    });
-    await expect(reconciler.advance(new AbortController().signal)).resolves.toMatchObject({
-      status: 'blocked',
-      reason: 'continuation-limit',
-      closureWireBytes: 0,
-    });
-    expect(apply).not.toHaveBeenCalled();
-  });
-
   it('pins one provider root and completes one active row across bounded slices', async () => {
     const fixture = await publishedFixture();
     const admission = admissionGate();
@@ -1264,194 +960,3 @@ describe('agent-profile System Record reconciler V1', () => {
     expect(loadInventoryObject).not.toHaveBeenCalled();
   });
 });
-
-async function publishedFixture() {
-  const fixture = await producerFixture();
-  const producer = createFixtureAgentProfileProducerV1({
-    networkId: NETWORK,
-    publicationDeployment: DEPLOYMENT,
-    peerSigner: fixture.peerSigner,
-    evmSigner: fixture.evmSigner,
-    store: fixture.store,
-    fence: () => undefined,
-    install: () => undefined,
-  });
-  await produce(producer, fixture.prepared, fixture.publication);
-  const rootArtifact = await fixture.store.resolve(
-    { type: 'root' },
-    new AbortController().signal,
-  );
-  if (rootArtifact === null) throw new Error('fixture root was not retained');
-  const rootEnvelope = parseCanonicalSignedSystemRecordRootDescriptorEnvelopeV1(
-    rootArtifact.canonicalBytes,
-  );
-  const inventory = fixture.store.snapshot().inventory;
-  if (inventory === null) throw new Error('fixture inventory was not retained');
-
-  return {
-    ...fixture,
-    rootEnvelope,
-    loadInventoryObject: async (
-      request: AgentProfileInventoryLoadRequestV1,
-      signal: AbortSignal,
-    ) => {
-      const stored = inventory.objects.get(request.objectDigest);
-      const objectKind = request.expectedKind;
-      const artifact = await fixture.store.resolve({
-        type: 'inventory-object',
-        rootDescriptorDigest: request.rootDescriptorDigest,
-        path: request.path,
-        objectKind,
-        objectDigest: request.objectDigest,
-      }, signal);
-      return artifact === null
-        ? {
-            outcome: 'rejected' as const,
-            wireBytes: 6,
-            rejection: 'not-found' as const,
-          }
-        : {
-          outcome: 'ok' as const,
-          objectKind,
-          canonicalBytes: artifact.canonicalBytes,
-          wireBytes: 4 + 128 + artifact.canonicalBytes.byteLength,
-        };
-    },
-  };
-}
-
-async function resignRootTotalRows(
-  fixture: Awaited<ReturnType<typeof publishedFixture>>,
-  totalRows: SignedSystemRecordRootDescriptorEnvelopeV1['object']['totalRows'],
-): Promise<SignedSystemRecordRootDescriptorEnvelopeV1> {
-  const object = Object.freeze({ ...fixture.rootEnvelope.object, totalRows });
-  return signRootDescriptor(fixture, object);
-}
-
-async function signRootDescriptor(
-  fixture: Awaited<ReturnType<typeof publishedFixture>>,
-  object: SignedSystemRecordRootDescriptorEnvelopeV1['object'],
-): Promise<SignedSystemRecordRootDescriptorEnvelopeV1> {
-  const objectDigest = computeSystemRecordRootDescriptorDigestV1(object);
-  const signature = await fixture.peerSigner.sign(buildSystemRecordProviderSignatureMessageV1(
-    object,
-    objectDigest,
-    fixture.peerSigner.peerId,
-  ));
-  return Object.freeze({
-    object,
-    objectDigest,
-    providerPeerId: fixture.peerSigner.peerId,
-    signatureSuite: 'ed25519-v1',
-    signature: Buffer.from(signature).toString('base64url'),
-  });
-}
-
-function receiver(
-  store: SystemRecordArtifactRepositoryV1,
-  consumeCandidate: Parameters<typeof createAgentProfileReceiverV1>[0]['consumeCandidate'],
-) {
-  return createAgentProfileReceiverV1({
-    networkId: NETWORK,
-    artifacts: store,
-    nowMs: () => PRODUCER_FIXTURE_NOW_MS,
-    verifyCurrentBundle: (_head, bundleBytes) => {
-      const { projectionBytes } = decodeOpaqueKaBundleV1(bundleBytes);
-      return Object.freeze({
-        canonicalProjectionBytes: Uint8Array.from(projectionBytes),
-        projectionQuads: Object.freeze(parseNQuads(new TextDecoder().decode(projectionBytes))),
-      });
-    },
-    consumeCandidate,
-  });
-}
-
-function receiverWithPreparation(
-  prepareActive: (
-    row: SystemRecordInventoryRowV1,
-    signal: AbortSignal,
-  ) => Promise<AgentProfilePreparedActiveV1>,
-): AgentProfileReceiverV1 {
-  const receiver: AgentProfileReceiverV1 = Object.freeze({
-    openPreparation(row) {
-      let released = false;
-      return Object.freeze({
-        prepare(_artifacts: SystemRecordArtifactRepositoryV1, signal: AbortSignal) {
-          if (released) throw new Error('test preparation is released');
-          return prepareActive(row, signal);
-        },
-        release() {
-          released = true;
-        },
-      });
-    },
-    prepareActive(row, _artifacts, signal) {
-      return prepareActive(row, signal);
-    },
-    async receiveActive(row, admittedContext, signal) {
-      const prepared = await prepareActive(row, signal);
-      signal.throwIfAborted();
-      return prepared.apply(admittedContext, signal);
-    },
-  });
-  return receiver;
-}
-
-function byteAdmission() {
-  return Object.freeze({
-    tryReserve() {
-      let released = false;
-      return Object.freeze({
-        release() {
-          released = true;
-        },
-        isReleased: () => released,
-      });
-    },
-  });
-}
-
-function admissionGate(
-  initiallyHeld = false,
-  nowMs: () => number = Date.now,
-): AgentProfileReconcileAdmissionV1 & {
-  stats(): { active: 0 | 1; peak: 0 | 1; acquisitions: number };
-  lastContext(): AgentProfileAdmittedSliceContextV1 | undefined;
-} {
-  let held = initiallyHeld;
-  let peak: 0 | 1 = held ? 1 : 0;
-  let acquisitions = 0;
-  let lastContext: AgentProfileAdmittedSliceContextV1 | undefined;
-  const contexts = new WeakMap<object, number>();
-  return Object.freeze({
-    tryAcquire() {
-      if (held) return null;
-      held = true;
-      peak = 1;
-      acquisitions += 1;
-      const context = Object.freeze(Object.create(null)) as AgentProfileAdmittedSliceContextV1;
-      contexts.set(context, nowMs() + 3_000);
-      lastContext = context;
-      let live = true;
-      return Object.freeze({
-        admittedContext: context,
-        release() {
-          if (!live) return;
-          live = false;
-          held = false;
-        },
-      });
-    },
-    inspectAdmittedContext(context) {
-      const admittedDeadlineMs = contexts.get(context);
-      if (admittedDeadlineMs === undefined) throw new Error('test admitted context is invalid');
-      return Object.freeze({ nowMs: nowMs(), admittedDeadlineMs });
-    },
-    stats() {
-      return { active: held ? 1 : 0, peak, acquisitions };
-    },
-    lastContext() {
-      return lastContext;
-    },
-  });
-}
