+++ packages/agent/vitest.unit.config.ts
@@ -107,6 +107,8 @@ export default defineConfig({
       "test/system-record-requester-v1.test.ts",
       "test/system-record-requester-wire-v1.test.ts",
       "test/system-record-reconcile-transport-v1.test.ts",
+      "test/system-record-reconcile-transport-integration-v1.test.ts",
+      "test/system-record-reconcile-closure-continuation-v1.test.ts",
       "test/sync-requester-bailout.test.ts",
       "test/swm-catchup-peer-selection.test.ts",
       "test/swm-fanout-peer-selection.test.ts",
